Arnage_Karting/Complete_Server_Django/rallys/requirements.txt
#Aquí las dependencias usadas
django_restframework
